Add length jvm code gen support

+++ bvm/ballerina-runtime/src/main/java/org/ballerinalang/jvm/values/ArrayValue.java
@@ -736,7 +736,7 @@ public class ArrayValue implements RefValue, CollectionValue {
         if (arrayElementType.getTag() == TypeTags.INT_TAG) {
             intValues = (long[]) newArrayInstance(Long.TYPE);
             for (int i = 0; i < this.size(); i++) {
-                intValues[i] = ((int) arrayValues[i]);
+                intValues[i] = ((long) arrayValues[i]);
             }
         }
